+++ bin/store/src/routers/root_store_router.rs
@@ -2,6 +2,7 @@ use crate::controllers::root_controller::{
     root_aggregation_filter, root_batch_delete_records, root_batch_insert_records,
     root_batch_update_records, root_create_record, root_delete_record, root_get_by_filter,
     root_get_by_id, root_search_suggestions, root_switch_account, root_update_record, root_upsert,
+    root_update_account_password,
 };
 use crate::middlewares::auth_middleware::Authentication;
 use crate::middlewares::session_middleware::SessionMiddleware;
@@ -32,6 +33,10 @@ pub fn configure_root_store_routes(cfg: &mut ServiceConfig) {
             .route(
                 "/{table}/filter/suggestions",
                 web::post().to(root_search_suggestions),
+            )
+            .route(
+                "/accounts/password/{account_id}",
+                web::put().to(root_update_account_password),
             ),
     );
 }